requirements.txt
cffi==1.14.3
cmarkgfm==0.4.2
Jinja2==2.11.2
MarkupSafe==1.1.1
pathlib==1.0.1
pycparser==2.20
python-frontmatter==0.5.0
PyYAML==5.1
six==1.15.0
